decision-log-backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg>=3.1.0
pgvector>=0.2.0
alembic>=1.12.0
anthropic>=0.7.0
langchain>=0.1.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
pyjwt>=2.8.0
python-multipart==0.0.6
httpx==0.25.2
sentry-sdk==1.39.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
ruff==0.1.8
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
APScheduler==3.10.4
